requirements.txt
asgiref==3.3.1
crispy==0.6.3
crispy-bootstrap5==0.3.1
cycler==0.10.0
Django==3.1.7
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-crispy-forms==1.11.1
django-js-asset==1.2.2
fabio==0.11.0
gunicorn==20.1.0
h5py==3.2.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
Pillow==8.1.2
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
pytz==2021.1
silx==0.14.0
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
